requirements.txt
fastapi~=0.115.14
uvicorn~=0.35.0
pydantic~=2.11.7
pymongo==4.13.2
google-generativeai==0.8.5
googlesearch-python==1.3.0